fix(auto-wrapup-timer): fix-data-types
